requirements.txt
streamlit
python-dotenv
tweepy
praw
openai
schedule
requests 